logic is working, readme has been updated

+++ bamazon.sql
@@ -8,7 +8,7 @@ CREATE TABLE products(
 id INT NOT NULL AUTO_INCREMENT,
 product_name VARCHAR(100) NOT NULL,
 department_name VARCHAR(100) NOT NULL,
-price INT(10) NOT NULL,
+price DECIMAL(10,2) NOT NULL,
 stock_quantity INT(100) NOT NULL,
 PRIMARY KEY (id)
 );
